Move FmtOptions methods to Fmt
